requirement.txt
asttokens==3.0.0
colorama==0.4.6
executing==2.2.0
peek-python==25.0.15
six==1.17.0
tomli==2.2.1
